% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dotplot_outlier.R
\name{analysis.dotplot.all}
\alias{analysis.dotplot.all}
\title{Generate Outlier Point Plots}
\usage{
analysis.dotplot.all(df, name = "dotplot.all", pdf = T)
}
\arguments{
\item{df}{\code{data.frame} with the data that should be analyzed for
outliers.}

\item{name}{\code{chr} name for the file that can be optionally outputted
via this function. The default name will be \code{dotplot.all}.}

\item{pdf}{\code{boolean} indicating whether the functions results should
be send to a PDF-file or whether they are intended for R-
internal use. Default is output to PDF (option \code{true}).}
}
\value{
\code{List} of dotplots (\code{ggplot}s) for \code{pdf = F} or a
         PDF file in the filesystem for \code{pdf = T}.
}
\description{
This function generates outlier point plots (for a description see
slide 117 of the DA1 lecture of WT 2015/16).
}
\details{
The method generates dotplots as univariate outlier detectors for all columns
of a given \code{data.frame} \code{df}. So via the function
\code{\link{analysis.dotplot}} it computes the distribution of values in each
given column via \code{\link{table}}. Based on the distribution the method
\code{\link{cpp_valueOccurrencesToPoints}} will generate visualizable points
representing each occurence in the form
\eqn{(x_i, y_{i,j})}{(x[i], y[i,j])}.
}
\examples{
KaggleHouse:::analysis.dotplot.all(iris, name = 'iris.dots', pdf = T)
}
\seealso{
\code{\link{analysis.dotplot}}

\code{\link{cpp_valueOccurrencesToPoints}}
}

