% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api.submit}
\alias{api.submit}
\title{Submit Results to Kaggle}
\usage{
api.submit(filepath, username, password, competition, message = "")
}
\arguments{
\item{filepath}{Path to the \code{csv} file with the predicted data to be
uploaded.}

\item{username}{Kaggle username to login to the Kaggle plattform.}

\item{password}{Kaggle password to login to the Kaggle plattform.}

\item{competition}{Name of the Kaggle competition the submission is intended
for.}

\item{message}{Additional message that describes the submission. Is
optional and by default empty.}
}
\description{
Upload a preformatted csv file with the predicted test results to Kaggle.
}
\details{
This method allows to upload the predicted output in csv format to the
Kaggle competition. It bases on the Kaggle CLI tool by floydwch which is a
Python CLI tool. This requires an existing Python installation as well as
having the Kaggle CLI tool installed (with Python and pip via
\code{pip install kaggle-cli}). With the requirements given this method
works as an R wrapper around this tool.
}
\examples{
 KaggleHouse:::api.submit(
   filepath = "sample_submission.csv",
   username = "kaggle_user",
   password = "abc123",
   competition = "house-prices-advanced-regression-techniques"
 )
}
\seealso{
Makes use of the Kaggle CLI tool by floydwch which can be found
         at \url{https://github.com/floydwch/kaggle-cli}.
}

