% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_bind}
\alias{cpp_bind}
\title{Combine two \code{NumericVector}s into a \code{NumericMatrix}}
\usage{
cpp_bind(a, b)
}
\arguments{
\item{a}{\code{NumericVector} representing the first column.}

\item{b}{\code{NumericVector} representing the second column.}
}
\value{
\code{NumericMatrix} being the combination of the two
        \code{NumericVector}s \code{a} and \code{b}.
}
\description{
This function combines two given \code{NumericVector}s into a
\code{NumericMatrix}. It thereby provides similar functionality as
\code{\link{base::cbind}}.
}

