% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_regex_selector_name}
\alias{cpp_regex_selector_name}
\title{Extract Selector Name From Variable Name}
\usage{
cpp_regex_selector_name(x)
}
\arguments{
\item{x}{\code{std::string} containing the selector name (usually in the
form \code{(containingObject)$(selector)}).}
}
\value{
\code{std::string} containing the extracted \code{selector} part.
        In case no selector can be found, the an empty string is returned.
}
\description{
This function extract the selector name (e.g. 'var') from a variable name
like 'data$var'.
}
\details{
This function basically takes variables names of the form
\code{(containingObject)$(selector)} as an input. With a regex command the
last selector - meaning the substring behind the last '$' - will be
extracted and returned. In case that the input parameter \code{x} does not
contain a selector, the empty string will be returned.
}
\examples{
  # The outcome should be simply 'SalePrice'
  KaggleHouse:::cpp_regex_selector_name(data_train_na$SalePrice)

}

