% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_rep_na_chr}
\alias{cpp_rep_na_chr}
\title{Remove \code{NA} from \code{CharacterVector}}
\usage{
cpp_rep_na_chr(xin, rep)
}
\arguments{
\item{xin}{\code{CharacterVector} potentially containing removable
\code{NA} values.}

\item{rep}{\code{std::string} value to replace \code{NA}s occuring in
\code{xin}.}
}
\value{
\code{CharacterVector} that is equal to \code{xin} except the
        \code{NA} values which have been replaced with \code{rep}.
}
\description{
This function removes \code{NA} values from \code{CharacterVector}s.
}
\details{
This function accepts a character input vector and then replaces each
contained \code{NA} value with the \code{std::string} specified in the
parameter \code{rep}.
}
\examples{
KaggleHouse:::cpp_rep_na_chr(c("a", "b", NA), "c")

}

