% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_rep_na_num}
\alias{cpp_rep_na_num}
\title{Remove \code{NA} from \code{NumericVector}}
\usage{
cpp_rep_na_num(xin, rep)
}
\arguments{
\item{xin}{\code{NumericVector} potentially containing removable \code{NA}
values.}

\item{rep}{\code{int} value to replace \code{NA}s occuring in \code{xin}.}
}
\value{
\code{NumericVector} that is equal to \code{xin} except the
        \code{NA} values which have been replaced with \code{rep}.
}
\description{
This function removes \code{NA} values from \code{NumericVector}s.
}
\details{
This function accepts a numeric input vector and then replaces each
contained \code{NA} value with the \code{int} specified in the parameter
\code{rep}.
}
\examples{
KaggleHouse:::cpp_rep_na_num(c(1, 2, NA), 3)

}

