% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_valueOccurrencesToPoints}
\alias{cpp_valueOccurrencesToPoints}
\title{Convert Value Occurences to Points}
\usage{
cpp_valueOccurrencesToPoints(x)
}
\arguments{
\item{x}{\code{data.frame} with the results of the \code{table} command on
a given vector. \code{x} should contain the counts per
characteristic of a specific feature.}
}
\value{
\code{List} including plottable point positions representing the
        counts of feature characteristics.
}
\description{
This function converts a \code{data.frame} of feature characteristic
occurences to plottable points. For that the characteristic value is used
as the x-value. To visualize multiple occurences, each of them is assigned
a slightly higher y-position, so that multiple occurences of one
characteristic will later on stack in the plotted graph.
}
\examples{
  val <- rep(1:5, c(10, 20, 30, 40, 50))
  df <- as.data.frame(table(val))
  KaggleHouse:::cpp_valueOccurrencesToPoints(df)
}

