% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_barplot.R
\name{create_barplot}
\alias{create_barplot}
\title{Create Bar Plot}
\usage{
create_barplot(data, col_name)
}
\arguments{
\item{data}{\code{vector} with the data of the column \code{col_name}.}

\item{col_name}{\code{character} string with the name of the column/feature
for which the bar plot is created.}
}
\value{
\code{ggplot2} bar plot for the feature \code{col.name} and data
        \code{data}.
}
\description{
This method acts as a wrapper around the \code{\link{ggplot2::ggplot}}.
}
\details{
This method creates a \code{\link{ggplot2::ggplot}} for the provided
\code{data}. The plotting function is preconfigured to create a bar plot
with \code{\link{ggplot2::theme_light}} and vertically aligned
characteristics on the x-axis (so that all can be displayed).
}
\examples{
 KaggleHouse:::create_barplot(data_train$MSZoning, "MS Zoning")
}

