% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate_data_views.R
\name{data.generate_data_views}
\alias{data.generate_data_views}
\title{Data view generation function}
\usage{
data.generate_data_views()
}
\description{
This function creates five different variables containing the names of features.
}
\details{
Calling this function is optional. It is useful when analysing different types
of features as e.g. continuous, discrete, nominal, ordinal and the target feature
Saleprice. All variables are bound the base environment.
}

