% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{echo}
\alias{echo}
\title{Set Cat Function to Echo variable}
\usage{
echo(..., file = "", sep = " ", fill = FALSE, labels = NULL,
  append = FALSE)
}
\description{
This function assigns the alias 'echo' to the cat-function.
}

