% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{echoln}
\alias{echoln}
\title{Write String to Output with appended Newline}
\usage{
echoln(str)
}
\arguments{
\item{str}{with the content to print to the output}
}
\description{
This function can be used to write a string \code{str} to the output with
an automatically appended newline.
}

