% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.checkInputParams}
\alias{feature.boruta.checkInputParams}
\title{Check Boruta Feature Selection Input Parameters}
\usage{
feature.boruta.checkInputParams(target, predictors, checkNA = F,
  verbose = F)
}
\arguments{
\item{target}{Response vector; factor for classification, numeric vector
for regression.}

\item{predictors}{\code{data.frame} with predictors.}

\item{checkNA}{\code{boolean} switch that decides whether the method will
conduct a check for \code{NA} values or not. Default is
\code{FALSE}.}

\item{verbose}{\code{boolean} switch that decides whether the error
output will provide more verbose information. Default is
\code{FALSE}.}
}
\description{
The method ensures that the parameters of the \code{\link{feature.boruta}}
function are valid.
}
\details{
This method analyses the \code{target} and \code{predictors} input variables
of the \code{\link{feature.boruta}} function. It conducts two mandatory
checks and one optional check. The first of the two mandatory checks
evaluates whether the correct data types are used. So the \code{target}
variable should be a \code{vector} while the \code{predictors} variable is
supposed to be a \code{data.frame}. As a second check it is ensured that
both variables have the same number of rows, which implies that the number of
considered observation is equal for both. If one of the checks fail,
execution will be aborted with an error explaining the reasons.
The optional check evaluates the presence of \code{NA} values as these are
not useable for many of the Boruta classifiers. In case either \code{target}
and/or \code{predictors} contain only one \code{NA} a warning is issued. If
the \code{verbose} switch is \code{TRUE}, the error will provide a list of
\code{predictors} columns which contain \code{NA} values.
}
\examples{
 KaggleHouse:::feature.boruta.checkInputParams(
   target = data_train_na$SalePrice, predictors = data_train_na[-81],
   checkNA = T, verbose = T
 )

}

