% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.fixNA}
\alias{feature.boruta.fixNA}
\title{Remove \code{NA} Containing Observations}
\usage{
feature.boruta.fixNA(target, predictors)
}
\arguments{
\item{target}{Response vector; factor for classification, numeric vector
for regression.}

\item{predictors}{\code{data.frame} with predictors.}
}
\value{
The \code{NA}-cleaned \code{data.frame} will be returned with two
        additional parameters:
        \item{target}{Containing the \code{NA}-cleaned target variable.}
        \item{predictors}{Containing the \code{NA}-cleaned predictors.}
}
\description{
Filter all observations from the \code{target} and \code{predictors} variable
that contain \code{NA} values to obtain a \code{NA} dataset usable by the
Boruta classifiers.
}
\details{
The method first carries out a reduced parameter check via
\code{\link{feature.boruta.checkInputParams}}. Afterwards both \code{target}
and \code{predictors} are combined to a unified \code{data.frame}. This
\code{data.frame} is then cleaned from \code{NA}s by the
\code{\link{na.omit}} function. Before splitting the \code{data.frame} into
a target and a predictor variable again, it is checked that at least one
observation prevailed. If that is not the case, the method is aborted with
an error.
}
\examples{
 KaggleHouse:::feature.boruta.fixNA(
   target = data_train_na$SalePrice, predictors = data_train_na[-81]
)

}

