% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.formula}
\alias{feature.boruta.formula}
\title{Create Formula Based on Boruta Selected Features}
\usage{
feature.boruta.formula(boruta, tentative = F)
}
\arguments{
\item{boruta}{\code{Boruta} object obtained by the execution of
\code{\link{feature.boruta}} method.}

\item{tentative}{\code{boolean} switch to decide whether tentative variables
should be part of the created formula. Default is
\code{FALSE}.}
}
\value{
\code{formula} object that represents the relationship between the
        predictors and the target variable which has been evaluated by the
        Boruta algorithm. The form will be \code{target ~ predictorVars}
        where \code{predictorVars} will only be confirmed (and tentative)
        variables.
}
\description{
The method creates a formula that puts a target value in relationship to
features selected by the Boruta algorithm.
}
\details{
This method evaluates the \code{Boruta} object. In case some other object is
provided to the function it will abort with an error. Given correct inputs
the function accesses the \code{target} property of the custom \code{Boruta}
object returned by \code{\link{feature.boruta}} and also obtains the
categorized list of the evaluated variables via
\code{\link{feature.boruta.variables}}. With these inputs a formula of the
structure \code{target ~ predictorVar1 + ...} is created which can then e.g.
be reused for regression.
}
\examples{
 boruta <- KaggleHouse:::feature.boruta(
   target = data_train_na$SalePrice, predictors = data_train_na[-81],
    fixNA = T, roughFix = T, verbose = T
 )

 KaggleHouse:::feature.boruta.formula(boruta, tentative = T)

}

