% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta}
\alias{feature.boruta}
\title{Boruta Feature Selection - Wrapper}
\usage{
feature.boruta(data = data_train_numeric_clean_imputed, recompute = F,
  desc = "")
}
\arguments{
\item{data}{\code{data.frame} containing the data on which the Boruta
feature selection should be executed. Will be used to feed
the \code{target} and \code{predictors} variables of
\code{\link{feature.boruta.comp}}. Defaults to
\code{data_train_numeric_clean_imputed}.}

\item{recompute}{\code{boolean} switch that determines if the Boruta feature
selection should be repeated when the results of a prior
run have been found. Recommended when parameters have
changed. Defaults to \code{FALSE}.}

\item{desc}{Additional comment that can be appended to the name of the
saved Boruta object. Can e.g. be used to store different
Boruta runs for different learners. Defaults to the empty
string.}
}
\description{
Convenience method that calls \code{\link{feature.boruta.comp}} with preset
parameters that are commonly used for learners in this package.
}
\details{
This method executes the packages Boruta wrapper
\code{\link{feature.boruta.comp}} with all parameters preset to fit the needs
of the package learners. Furthermore it stores each Boruta object after the
computation to the \code{output/feature_selection} directory. Before a new
computation is started the directory is checked for the existance of an
already computed Boruta object. If one is available and the \code{recompute}
flag is \code{FALSE} the previously computed object is loaded and used.
Finally the method binds all selected features (confirmed and tentative ones)
to the \code{.GlobalEnv} as the \code{features_boruta} variable.
}
\examples{
 KaggleHouse:::feature_boruta()
 KaggleHouse:::feature_boruta(recompute = T)
 KaggleHouse:::feature_boruta(recompute = T, desc = "_test_run_")
}

