% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.report}
\alias{feature.boruta.report}
\title{Create PDF and Text Reports About Selected Boruta Features}
\usage{
feature.boruta.report(boruta)
}
\arguments{
\item{boruta}{The \code{Boruta} object for which a report should be
generated.}
}
\description{
This methods provides a PDF with the major info graphics related to Boruta
feature selection, as well as a text file with all confirmed, tentative and
rejected variables including the formula generated from them.
}
\details{
In a first step this method will generate two plots for the given
\code{Boruta} object: One plot that will show boxplots for all features,
as well as their importance and their selection status (e.g. a red feature
has been discared, a green one confirmed). Additionally the imputation
history is added, which shows the importance (and acceptance status) of each
feature over the time. These two graphics are combined into one PDF file.
Since it is versioned with time information, each call to this method will
automatically generate a new report. The time stamp thus allows to version
different \code{Boruta} objects.
In case the \code{Boruta} object has been generated with
\code{\link{feature.boruta.comp}} and contains the \code{Confirmed} (etc.)
variables an additional text file is generated and filled with all the
information regarding those features. If the \code{formula} switch has
been activated the generated formula will also be added to that text file.
Similar to the PDF file the txt file is versioned via timestamps as part
of the file name.
}
\examples{
 boruta <- KaggleHouse:::feature.boruta(
   target = data_train_na$SalePrice, predictors = data_train_na[-81],
    fixNA = T, roughFix = F, variables = T, selected = T, formula = T,
    verbose = T
 )
 KaggleHouse:::feature.boruta.report(boruta)

}

