% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.selected}
\alias{feature.boruta.selected}
\title{Provide all Non-Rejected Boruta Features}
\usage{
feature.boruta.selected(boruta)
}
\arguments{
\item{boruta}{The \code{Boruta} object to which the \code{Selected} variable
should be attached.}
}
\value{
\code{Boruta} object with an attached \code{Selected} variable
        containing the confirmed and tentative features.
}
\description{
This method combines the features that have either been confirmed or have
been marked as tentative in the Boruta algorithm into one variable.
}
\details{
This method usually is not meant to be applied by a package user. In case it
is still used to obtain the confirmed and tentative variables of a
\code{Boruta} object it has to be ensured that it has the variables
\code{Confirmed} and \code{Tentative}. These can be obtained via the
\code{variables} switch of the \code{\link{feature.boruta.comp}} method that
has to be \code{TRUE}.
Then the features contained in both variables will be combined into a
\code{Selected} variable. Finally the method binds the \code{Selected}
variable to the \code{.GlobalEnv} as \code{features_boruta}.
}
\examples{
 boruta <- KaggleHouse:::feature.boruta(
   target = data_train_na$SalePrice, predictors = data_train_na[-81],
    fixNA = T, roughFix = F, variables = T, verbose = T
 )
 KaggleHouse:::feature.boruta.selected(boruta)

}

