% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.tentative}
\alias{feature.boruta.tentative}
\title{Check For Tentative Variables}
\usage{
feature.boruta.tentative(boruta)
}
\arguments{
\item{boruta}{\code{Boruta} object obtained by the execution of
\code{\link{feature.boruta}} method.}
}
\value{
\code{boolean} value that is \code{TRUE} as soon as one tentative
        variable is found. \code{FALSE} otherwise.
}
\description{
The method checks a \code{Boruta} object for the decision status by checking
if some variables are still tentative.
}
\details{
This method evaluates the \code{Boruta} object. In case some other object is
provided to the function it will abort with an error. All correctly assigned
objects will then be checked for variables that still have the tentative
status. If at least one is still tentative a \code{TRUE} will be returned.
}
\examples{
 boruta <- KaggleHouse:::feature.boruta(
   target = data_train_na$SalePrice, predictors = data_train_na[-81],
    fixNA = T, roughFix = T, verbose = T
 )

 KaggleHouse:::feature.boruta.tentative(boruta)

}

