% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.boruta.R
\name{feature.boruta.variables}
\alias{feature.boruta.variables}
\title{Obtain List of Confirmed, Tentative and Rejected Variables}
\usage{
feature.boruta.variables(boruta)
}
\arguments{
\item{boruta}{\code{Boruta} object obtained by the execution of
\code{\link{feature.boruta.comp}} method.}
}
\value{
A list with the three entries:
         \item{Confirmed}{Vector of confirmed variables.}
         \item{Tentative}{Vector of tentative variables.}
         \item{Rejected}{Vector of rejected variables.}
}
\description{
The method obtains a list of three vectors containing confirmed, tentative
and rejected variables.
}
\details{
This method evaluates the \code{Boruta} object. In case some other object is
provided to the function it will abort with an error. From correctly provided
\code{Boruta} objects a list of all evaluated variable names will be
obtained. Based on the \code{finalDecision} variable of the \code{Boruta}
object the variable names will be split into three vectors representing
confirmed, tentative and rejected variables. These vectors are combined in a
list and are returned as such.
}
\examples{
 boruta <- KaggleHouse:::feature.boruta(
   target = data_train_na$SalePrice, predictors = data_train_na[-81],
    fixNA = T, roughFix = T, verbose = T
 )

 KaggleHouse:::feature.boruta.variables(boruta)

}

