% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.lasso.R
\name{feature.lasso}
\alias{feature.lasso}
\title{Lasso feature selection}
\usage{
feature.lasso(data = data_train_numeric_clean_imputed)
}
\arguments{
\item{data}{Input data}
}
\description{
When executed an optimal set of features according to the lambda value
that minimizes the loss is bound to a global variable \code{features_lasso}.
}
\details{
Default dataset is \code{data_train_numeric_clean_imputed}. The family
is \code{gaussian} and for measuring the goodness \code{auc} is chosen.
}
\examples{
     feature.lasso(data=BostonHousing)

}

