% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_barplot.R
\name{general_barplot}
\alias{general_barplot}
\title{Generate Bar Plots for Dataset Analysis}
\usage{
general_barplot(data = data_train, data_cols = c(iowa.houses.nominal.vars,
  iowa.houses.ordinal.vars), name = "barplots", pdf = T)
}
\arguments{
\item{data}{\code{data.frame} containing the data for which the bar
plots should be created.}

\item{data_cols}{\code{vector} of column names containing nominal and ordinal
data (for which the bar plots should be generated).}

\item{name}{\code{character} string that is used as the filename.}

\item{pdf}{\code{boolean} switch to either export the graphics to a PDF
file or print them on screen. Defaults to \code{TRUE}.}
}
\description{
This method generates bar plots for the nominal and ordinal variables of a
given dataset to visualize their value distributions.
}
\details{
This method is used to visualize the distribution of nominal and ordinal
features in a given dataset \code{data}. As not automated detection of these
features is conducted, the columns containing such data are given by the
\code{data_cols} parameter. For each of the features a bar plot is created
via the \code{\link{create_barplot}} method.
If the \code{pdf} switch is activated (\code{switch == TRUE}) the bar plots
will be exported to a PDF file. Here the filename is determined by the
\code{name} attribute. Each PDF page will then contain up to two bar plots.
Otherwise the bar plots will be shown on the screen.
}
\examples{
 # Create vectors containing the column names of columns with oridinal or
 # nominal features.
 KaggleHouse:::data.generate_data_views()
 # Generate PDF bar plots.
 KaggleHouse:::general_barplot()

}

