% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_hist.R
\name{general_hist}
\alias{general_hist}
\title{Create Histogram of Numeric Data (to visualize Distribution)}
\usage{
general_hist(x, name, pdf = T)
}
\arguments{
\item{x}{dataset to be analysed}
}
\description{
The function creates histograms for all columns/variables with numeric
variables in the given dataset \code{x} (at least this is the default
behaviour) and saves them to a PDF file.
Creating the histograms helps to get a better understanding of the
distribution of the values. As the visualized distributions might ressemble
the normal distribution, the p-value is plotted below the histograms as an
additional indicator for normality.
}

