% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_plot.R
\name{general_plot}
\alias{general_plot}
\title{Create Scatterplot of Data (to visualize Correlations)}
\usage{
general_plot(x, name, threshold = 0.7, pdf = T, o = T,
  cor.met = "pearson")
}
\arguments{
\item{x}{Dataset to be analysed.}

\item{name}{Name that will be used in case a PDF will be produced.}

\item{threshold}{The minimum correlation that two features must have to be
considered further. If a feature does not have a correlation
greater than the threshold with at least two other features
it will be excluded from the correlation plot. Defaults to
a value of 0.7.}

\item{pdf}{Switch variable that decides whether the plot will be output to
PDF or not. Defaults to PDF output.}

\item{o}{\code{Boolean} switch to decide whether the method is allowed to
print progress output to \code{stdout}. Defaults to \code{true}.}

\item{cor.met}{Methodology used to compute the correlation. Can be "pearson",
"spearman" oder "kendall". Defaults to "pearson".}
}
\description{
The function creates a scatterplot of the complete dataset given by the
variable and saves it in a PDF file.
The lower panel makes use of \code{panel.smooth} to create the usual
scatterplots for pairs of two variables but also adds some kind of a
'regression line' to the plot to increase the understandability.
The upper panel uses a custom panel that shows of the numeric correlation
values (highlighted with different font sizes and colors).
}

