% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_qq.R
\name{general_qq}
\alias{general_qq}
\title{Create a QQ-Plot of Numeric Data (to Analyse the Normality Assumption)}
\usage{
general_qq(x, name, pdf = T)
}
\arguments{
\item{x}{dataset to be analysed}
}
\description{
This function takes a dataset \code{x} - which by default will consist of all
numeric variables of the Tripadvisor hotel dataset - and creates a QQ-plot
for each of the contained columns/variables.
To improve the ease of understanding whether the given data is rather normal
distributed or not (theoretical and practical quantiles are similar)
additional QQ-Lines are added.
}

