% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_summary.R
\name{general_summary_detail}
\alias{general_summary_detail}
\title{Data Summary - Get a Summary of the Data}
\usage{
general_summary_detail(x, name)
}
\arguments{
\item{x}{dataset to be analysed}
}
\description{
This function executes the \code{summary}-function on inputted datasets
\code{x}. The results are written to a csv-file for later use.
}

