% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_train.R
\name{imputation.train}
\alias{imputation.train}
\title{Impute training data}
\usage{
imputation.train(data = data_train_numeric_clean)
}
\arguments{
\item{data}{Input data which is set by default to \code{data_train_numeric_clean}}
}
\description{
Imputation function for the training dataset using package \code{mice} and manual
substitution to remove missing values.
}
\details{
Since some features are not missing at random imputation is not an appropriate approach.
Therefore the features \code{MasVnrArea}, \code{LotFrontage}, \code{Electrical} and
\code{GarafeYrBlt} are manually imputed. The remaining features that still contain
at least one missing value are solely imputed using \code{mice} with 50 maximum
iterations with seed 500 using predictive-mean matching. After the execution one
can access the variable \code{data_train_numeric_clean_imputed}.
}

