% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.deeplearning.R
\name{learner.deeplearning}
\alias{learner.deeplearning}
\title{Deep Learning with h2o tuned by mlr}
\usage{
learner.deeplearning(data = data_train_numeric_clean_imputed)
}
\arguments{
\item{data}{\code{data.frame} containing the data that should be used for
deep learning. For optimal results an imputed and cleaned data
set should be provided. The default data set is the KaggleHouse
train data set after imputation and conversion into a numeric
\code{data.frame}.}
}
\description{
This method performs parameter tuning and feature selection on the provided
\code{data} data set.
}
\details{
This method trains an ANN based on the \code{\link{h2o}} package (more
specificially the \code{\link{h2o.deeplearning}}) function. For the
training it uses the given \code{data} and tries to adjust the \code{hidden}
(number of hidden layers) and the \code{rate} (learning rate) parameters.
On top feature selection will be performed to only keep those features
actually contributing to the model.
The final results will be saved to the
\code{learner.deeplearning_result.RData} file. This way they can later be
reused to extract the optimal parameters for a deeplearning ANN.
}
\examples{
 KaggleHouse:::learner.deeplearning(data_train_numeric_clean_imputed)

}
\references{
A. Candel, J. Lanford, E. LeDell, V. Parmar, A. Arora (2015).
 Deep Learning with H2O (\emph{Third Edit.})
 Publisher: \emph{H2O.ai, Inc.}
 URL: \url{http://h2o.gitbooks.io/deep-learning/}

 S. Aiello, T. Kraljevic and P. Maj (2016).
 h2o: R Interface for H2O
 URL: \url{http://www.h2o.ai/}
}

