% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.stacked.R
\name{learner.stacked}
\alias{learner.stacked}
\title{Stacked learner}
\usage{
learner.stacked(input_train = data_train_numeric_clean_imputed,
  input_test = data_test_numeric_clean_imputed)
}
\arguments{
\item{input_train}{Input data which is by default set to \code{data_train_numeric_clean_imputed}}

\item{input_test}{Input data which is by default set to \code{data_test_numeric_clean_imputed}}
}
\description{
Combines the parameter sets from the tuning of the basic learners to predict the final Saleprice
using linear regression.
}
\details{
Uses the \code{features_boruta} to select only features that are considered important by \code{feature.boruta()}.
This stacked learner uses xgboost, lasso with the tuned parameters and deeplearning with 10 hidden layers each
containing 300 nodes as basic learners. At last linear regression is used to predict the Saleprice using
all three predictions. The result is stored in \code{final_submission_stacked_learner.csv} and can be directly
uploaded to kaggle.
}
\examples{
 KaggleHouse:::learner.stacked()

}

