% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_against_var.R
\name{plot_against_var}
\alias{plot_against_var}
\title{2D plot generator}
\usage{
plot_against_var(df, var, pdf = TRUE)
}
\arguments{
\item{df}{Dataframe containing all features to which \code{var should be plotted}}

\item{var}{Vector which is the variable that is plotted to each variable in \code{df}}

\item{pdf}{Boolean variable stating whether to create a pdf or not}
}
\description{
Generates a set of two dimensional plots where one variable is plotted against all
other variables.
}
\details{
All two dimensional plots are saved in one pdf file named
\code{understand_data.data_against} using ggplot. This function is typically used
to find correlations to the target variable.
}

