% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_transform_data.R
\name{prepare.transform_data}
\alias{prepare.transform_data}
\title{Data transformation function}
\usage{
prepare.transform_data(data, na.col)
}
\arguments{
\item{data}{Input data which contains missing values}

\item{na.col}{Set of column names which contain at least one
missing value that is then substituted with the character
"None"}
}
\description{
Substitutes missing values of a specific set of columns with the
character "None".
}
\details{
Especially in the given dataset missing values often yielded
semantical meaning. To generate an additional class when transformed
into numerical values this function is used to substitutes those
missing values not at random with a character "None" that should
get transformed into an numeric value instead of NA.
}
\examples{
 KaggleHouse:::prepare.transform_data(data.frame(id = c(1,2,NA)), "id")

}

