% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{preprocess.generate_cleaned_data}
\alias{preprocess.generate_cleaned_data}
\title{Cleaning data function}
\usage{
preprocess.generate_cleaned_data(data, rows, cols)
}
\arguments{
\item{data}{Input data that is supposed to be reduced in number of columns and rows}

\item{rows}{Rows that should be removed from \code{data}}

\item{cols}{Columns that should be removed from \code{data}}
}
\value{
\code{data} with removed rows and columns
}
\description{
Main purpose is to remove a set of rows and columns from a given dataset.
}
\details{
After checking for valid input parameter first rows and then
columns are removed from the \code{data}. The cleaned dataset is globally available under
the name of the input dataset with "_clean" as prefix.
}

