% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_conversion.R
\name{rda.conversion.checkRawData}
\alias{rda.conversion.checkRawData}
\title{Get List of RAW data files.}
\usage{
rda.conversion.checkRawData(directory = "inst/extdata")
}
\arguments{
\item{directory}{The directory in which the RAW files should be searched.}
}
\value{
\code{List} of files present in the specified \code{directory}.
}
\description{
The function checks the given \code{directory} for RAW data files.
}
\details{
First the presented \code{directory} is checked for existance. In case it
cannot be found a \code{warning} will be thrown. If it is present, a list of
all files in the directory will be created and returned.
These files will then be considered as candidates for being loaded as
package data.
}

