% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_conversion.R
\name{rda.conversion.ensureDataDir}
\alias{rda.conversion.ensureDataDir}
\title{Ensure Existance of Data Directory}
\usage{
rda.conversion.ensureDataDir()
}
\description{
The function checks the existance of the \code{data} directory and creates it
in case it is necessary.
}
\details{
The function ensures the presence of a \code{data} directory via the utility
helper function \code{\link{util.ensure.dir}}. This method then checks the
availability of the folder and creates a new folder in case it should be
necessary.
}
\seealso{
\code{\link{util.ensure.dir}}
}

