% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_conversion.R
\name{rda.conversion.loadDataFile}
\alias{rda.conversion.loadDataFile}
\title{Load data from RAW data file}
\usage{
rda.conversion.loadDataFile(directory, file, header = T, sep = ",",
  dec = ".", quote = "\\"")
}
\arguments{
\item{directory}{The directory in which the target \code{file} is located.}

\item{file}{The file that should be loaded by this function.}

\item{header}{\code{boolean} specifying whether the to-read-CSV has a header
row or not. Defaults to \code{TRUE}.}

\item{sep}{Character specifying the \code{chr} separating two data entries.
Defaults to \code{,}.}

\item{dec}{Character specifying the decimal split \code{chr}. Defaults to
\code{.}.}

\item{quote}{Set of quoting characters. Defaults to \code{""}.}
}
\value{
\code{data.frame} containing the data present in the specified
         \code{file}.
}
\description{
A function that loads raw data from a CSV file to a local variable.
}
\details{
The function first checks the existance of the given \code{directory} and
\code{file}. In case both exist, the specified \code{file} will be loaded
considering the loading parameters \code{header, sep, dec} and \code{quote}.
}
\examples{
 # Assuming the existance of directory 'dir' and file 'data.csv':
 KaggleHouse:::rda.conversion.loadFile(
     'dir', 'data.csv', header = T, sep = ";", dec = ".", quote = '\\"'
 )
}

