% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_conversion.R
\name{rda.conversion.saveAsRDA}
\alias{rda.conversion.saveAsRDA}
\title{Convert Loaded Data to RDA File.}
\usage{
rda.conversion.saveAsRDA(dat, name)
}
\arguments{
\item{dat}{\code{data.frame} containing the data loaded from a raw data/CSV-
file.}

\item{name}{Name assigned to \code{dat} by the filename of it's raw file.}
}
\description{
This function converts a loaded dataset \code{dat} into an RDA file.
}
\details{
This function takes a dataset \code{dat} (typically previously loaded via
\code{\link{rda.conversion.loadDataFile}}) and the \code{name} it comes with.
Based on this it first checks the \code{data} directory for an RDA file with
the same name (indicating that the data in question has already been
converted). Whenever conversion is still required, the input data \code{dat}
will be bound to the \code{.GlobalEnv} under the label of \code{name}. This
allows to store the variable content to an RDA file via
\code{\link{base::save}}. Finally \code{.GlobalEnv} will be cleaned from
\code{dat} again.
}
\examples{
# Assuming the existance of directory 'inst/extdata' and file 'data.csv':
imp <- KaggleHouse:::rda.conversion.loadDataFile('inst/extdata', 'data.csv')
KaggleHouse:::rda.saveAsRDA(imp, 'data.csv')
}
\seealso{
\code{\link{rda.conversion.loadDataFile}}
}

