% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_descriptive}
\alias{run_descriptive}
\title{Descriptive analysis}
\usage{
run_descriptive(train = data_train,
  train_no_na = data_train_numeric_clean_imputed)
}
\arguments{
\item{train}{Training dataset}

\item{train_no_na}{Training dataset only containing numerical values and no missing data}
}
\description{
Mainly used to output plots that are used to get a general understanding of the given data.
}
\details{
The plots and text documents are saved under the /output directory. Besides generating a
textual summary of the whole training dataset this function also generates histograms, barplots,
scatterplots and qq-plots.
}

