% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_generate_data}
\alias{run_generate_data}
\title{Function for generating datasets}
\usage{
run_generate_data()
}
\description{
In general the first function that should be executed when the package is loaded.
Generates the two dataset \code{data_train_numeric_clean_imputed} and \code{data_test_numeric_clean_imputed}.
}
\details{
First the function generates data views for e.g. distinguishing categorical and ordinal variables. Next the
data is cleaned by removing unnecessary columns and rows. After that imputation takes place by manually
and automatically(mice with pmm) substituting missing values.
}

