% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_learner}
\alias{run_learner}
\title{Main stacked learner}
\usage{
run_learner(train = data_train_numeric_clean_imputed,
  test = data_test_numeric_clean)
}
\arguments{
\item{train}{Cleaned, imputed and numeric training dataset}

\item{test}{Cleaned, imputed and numeric test dataset}
}
\description{
Executes the stacked learner on the train and test dataset
}
\details{
Wrapper function that is exported to the user of the package.
}

