% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.contain_na}
\alias{util.contain_na}
\title{Check Data Columns for \code{NA} Values}
\usage{
util.contain_na(data)
}
\arguments{
\item{data}{\code{data.frame} in which each column should be checked for the
presence of \code{NA} values.}
}
\value{
\code{Named logi} (being a \code{logical} vector also containing the
        column name for each \code{logical} value) indicating the presence
        of \code{NA} values for each column of \code{data}.
}
\description{
This method returns a \code{logical} vector that indicates the presence of
\code{NA} values in every column of \code{data}.
}
\details{
This method takes a \code{data.frame} as an input and then checks for each
column if it contains any \code{NA} values. The results get stored in a
\code{logical} vector (always together with the column name).
}
\examples{
 KaggleHouse:::util.contain_na(data_train)

}

