% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.ensure.dir}
\alias{util.ensure.dir}
\title{Ensure Existance of Directory}
\usage{
util.ensure.dir(dir.name)
}
\arguments{
\item{dir.name}{\code{character} string of the name of the directory which
should be checked for existance and added if missing.}
}
\description{
This method ensures that the directory given as a parameter exists.
}
\details{
This method takes a \code{character} string as input which denotes a
directory path. It first checks whether the given directory already exists
on the filesystem. If this is not the case it will be created.
}
\examples{
 KaggleHouse:::util.ensure.dir("output")
}

