% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.generate.submit}
\alias{util.generate.submit}
\title{Generate a Kaggle-submitable CSV from the Predicted Data}
\usage{
util.generate.submit(data, name)
}
\arguments{
\item{data}{\code{data.frame} containing the data needed for a submission of
predictions to the Kaggle platform.}

\item{name}{\code{character} string that represents the name of the CSV file
to which the submission data will be stored.}
}
\description{
This method generates a CSV file submitable to Kaggle from a
\code{data.frame} containing the relevant data (incl. the predictions).
}
\details{
This method takes a \code{data.frame} that already contains all relevant
data for a prediction submission to the Kaggle platform. E.g. in case of the
"House Prices: Advanced Regression Techniques" challenge this
\code{data.frame} would have to include one column with the ID and one with
the predicted sale price of a house.
These data is then written to a CSV file via a preconfigured call to the
\code{\link{write.csv}} function to a file with the name given in the
\code{name} parameter.
}

