% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.list_to_df}
\alias{util.list_to_df}
\title{Convert List of Matrices to \code{data.frame}}
\usage{
util.list_to_df(lst)
}
\arguments{
\item{lst}{List with matrices.}
}
\value{
\code{data.frame} being the concatenated version of all matrices in
        \code{list} \code{lst}.
}
\description{
Transforms a list of matrices (with the same columns ['objects', 'counts'])
into a concatenated \code{data.frame}.
}
\details{
This method first uses the \code{\link{util.list_to_matrix}} method to
convert the given list of matrices to the concatenated version of all
matrices stored in a single matrix. Afterwards the derived matrix is
transformed into the \code{data.frame} returned by the function.
}

