% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.mse}
\alias{util.mse}
\title{Compute the MSE between two \code{data.frame}s}
\usage{
util.mse(data_predicted, data_original)
}
\arguments{
\item{data_predicted}{First \code{data.frame} for the MSE computation.}

\item{data_original}{Second \code{data.frame} for the MSE computation.}
}
\description{
This method is used to compute the mean squared error (MSE) between two
\code{data.frame}s.
}

