% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.na.rm}
\alias{util.na.rm}
\title{Clear Matrix of all \code{NA} rows.}
\usage{
util.na.rm(x, fun = any)
}
\arguments{
\item{x}{with the matrix to clear from \code{NA} values}

\item{fun}{with the function to determine the number of allowed \code{NA}s}
}
\value{
Matrix with cleaned rows
}
\description{
Function to remove rows from a given matrix \code{m} which contain a
specified amount of \code{NA} values. The default is that a row is removed
from the matrix as soon as one \code{NA} value is included. But it is also
possible to configure the method to remove rows only when all values are
\code{NA}.
}

