% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.panel.cor}
\alias{util.panel.cor}
\title{Define a Scatterplot Panel to Show Correlation Coefficients}
\usage{
util.panel.cor(x, y, digits = 2, cex.cor, ...)
}
\arguments{
\item{x}{first dataset}

\item{y}{second dataset}

\item{digits}{number of digits after the decimal separator}

\item{cex.cor}{desired font/text size}

\item{...}{further arguments}
}
\description{
This function calculates the correlation between two datasets \code{x} and
\code{y} and writes the textual representation into the corresponding field
of the scatterplot panel.
Depending on the derived value different font sizes and color schemes are
applied.
}

