% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.remove.fileextension}
\alias{util.remove.fileextension}
\title{Remove File-Extension from Filepath}
\usage{
util.remove.fileextension(filePath)
}
\arguments{
\item{filePath}{\code{character} string representing a filepath like
\code{folder/file.ext}.}
}
\value{
\code{character} string representing path in parameter
         \code{filePath} without the fileextension.
}
\description{
This function takes a filepath including the extension, removes the extension
and returns the remaining path.
}
\details{
This function accepts a filepath of the structure \code{folder/file.ext}. It
determines the last point in the character string via a RegEx operation. The
underlying logic is that most file extensions will be separated from the
remaining path by that point. Then only the part of the filepath up to that
point is extracted and kept. So the resulting filepath should look like
\code{folder/file}.
}
\examples{
 # Input is 'data.csv' and output should be 'data'.
 KaggleHouse:::util.remove.fileextension("data.csv")

}

