% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.write_csv}
\alias{util.write_csv}
\title{Write data to a CSV-file.}
\usage{
util.write_csv(data, name, ...)
}
\arguments{
\item{data}{being a matrix or data-frame containing the writable data}

\item{name}{being a string providing the filename}

\item{...}{additional arguments to write.csv2}
}
\description{
Wrapper function around \code{write.table} that writes a given set of data
to a file with the name \code{name}. Predefined separator is \code{;} and the
used file-encoding is \code{utf-8}.
}

